* master: (201 commits)
  modules: bunch of READMEs regenerated to be sync'ed with docbook
  modules: coherent naming of Statistics section in docbook
  sl: regenerated the readme
  sl: documented exported statistics
  core/tcp: define option TCP_CLONE_RCVBUF to turn off/on tcp rcv buffer cloning
  pdt(k): added rpc command pdt.reload
  pdt(k): MI commands can be disabled via define PDT_NO_MI
  pdt(k): updates to documentation
  pdt: refactored prefix-domain translation
  tcp: fix for ENOTCONN on newer FreeBSDs
  dialplan: added fnmatch (2) matching operator
  core/tcp: init max_tls_connections in tcp options check
  core/tcp: proper init of parameter tls_max_connections
  textopsx: added fnmatch(value, expr, flags) function
  core: added function to help fixup of PVE, string and expressions
  registrar(k): maintaining one contact per AoR has priority over max_contacts
  lib/srdb1: Enable non-pooled database connections to be defined.
  modules_k/presence: Added additional return value to pres_auth_status() for polite-block
  tls: new parameter 'renegotiation' to enable/disable client renegotiation
  tls: use pkg-config to get cflags and libs for compilation
  ...

+++ modules_k/pv/pv_xavp.c
@@ -42,11 +42,11 @@ int pv_xavp_get_value(struct sip_msg *msg, pv_param_t *param,
 			return pv_get_strval(msg, param, res, &avp->val.v.s);
 		break;
 		case SR_XTYPE_TIME:
-			if(snprintf(_pv_xavp_buf, 128, "%lu", avp->val.v.t)<0)
+			if(snprintf(_pv_xavp_buf, 128, "%lu", (long unsigned)avp->val.v.t)<0)
 				return pv_get_null(msg, param, res);
 		break;
 		case SR_XTYPE_LONG:
-			if(snprintf(_pv_xavp_buf, 128, "%ld", avp->val.v.l)<0)
+			if(snprintf(_pv_xavp_buf, 128, "%ld", (long unsigned)avp->val.v.l)<0)
 				return pv_get_null(msg, param, res);
 		break;
 		case SR_XTYPE_LLONG:
